cw support, wasm hook support for memo, xcvm and assets network id compatibility (#3957)
I have made some changes to use more enums so it is easier to handle
various cases.
It usual patter in Rust and Substrate and CW.

Required for merge:
- [ ] `pr-workflow-check / draft-release-check` is ✅ success
- Other rules GitHub shows you, or can be read in
[configuration](../terraform/github.com/branches.tf)

Makes review faster:
- [x] PR title is my best effort to provide summary of changes and has
clear text to be part of release notes
- [x] I marked PR by `misc` label if it should not be in release notes
- [x] Linked Zenhub/Github/Slack/etc reference if one exists
https://github.com/ComposableFi/composable/issues/3953
- [x] I was clear on what type of deployment required to release my
changes (node, runtime, contract, indexer, on chain operation, frontend,
infrastructure) if any in PR title or description
- [x] Added reviewer into `Reviewers`
- [x] I tagged(`@`) or used other form of notification of one person who
I think can handle best review of this PR
- [ ] I have proved that PR has no general regressions of relevant
features and processes required to release into production
- [x] Any dependency updates made, was done according guides from
relevant dependency
- Clicking all checkboxes 
- Adding detailed description of changes when it feels appropriate (for
example when PR is big)

+++ code/parachain/frame/composable-traits/src/xcm/memo.rs
@@ -15,12 +15,12 @@ use crate::prelude::*;
 	Debug,
 )]
 pub struct ChainInfo {
-	pub chain_id: u128,
+	pub chain_id: u32,
 	pub order: u8,
 	pub channel_id: u64,        //for packet or memo
 	pub timestamp: Option<u64>, //for packet
 	pub height: Option<u64>,    //for memo packet message forwarding
-	pub retries: Option<u64>,   //for memo packet message forwarding
+	pub retries: Option<u8>,    //for memo packet message forwarding
 	pub timeout: Option<u64>,   //for memo packet message forwarding
 	pub is_substrate_ibc: bool,
 	pub is_substrate_xcm: bool,
